refactor(keys): move public key back to keys module
