+++ src/keys/mod.rs
@@ -1,11 +1,58 @@
 use crate::error::Error;
-use ed25519_dalek::{ExpandedSecretKey, PublicKey, SecretKey};
-use k256::ecdsa::{
-    signature::{Signer as EcdsaSigner},
-    Signature as EcdsaSignature, SigningKey,
-};
+use ed25519_dalek::{ExpandedSecretKey, SecretKey};
+use k256::ecdsa::{VerifyingKey, signature::{Verifier as EcdsaVerifier}};
+use k256::ecdsa::{Signature as EcdsaSignature, SigningKey, signature::{Signer as EcdsaSigner}};
 use serde::{Deserialize, Serialize};
 use zeroize::Zeroize;
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct PublicKey {
+    public_key: Vec<u8>
+}
+
+impl PublicKey {
+    pub fn new(key: Vec<u8>) -> Self {
+        PublicKey {public_key: key.to_vec()}
+    }
+
+    pub fn key(&self) -> Vec<u8> {
+        self.public_key.clone()
+    }
+
+    pub fn verify_ed(&self, msg: &[u8], sig: &[u8]) -> bool {
+        if let Ok(key) = ed25519_dalek::PublicKey::from_bytes(&self.key()) {
+            use arrayref::array_ref;
+            if sig.len() != 64 {
+                return false;
+            }
+            let sig = ed25519_dalek::Signature::from(array_ref!(sig, 0, 64).to_owned());
+            match key.verify(msg, &sig) {
+                Ok(()) => true,
+                Err(_) => false,
+            }
+        } else {
+            false
+        }
+    }
+
+    pub fn verify_ecdsa(&self, msg: &[u8], sig: &[u8]) -> bool {
+        match VerifyingKey::from_sec1_bytes(&self.key()) {
+            Ok(k) => {
+                use k256::ecdsa::Signature;
+                use std::convert::TryFrom;
+                if let Ok(sig) = Signature::try_from(sig) {
+                    match k.verify(msg, &sig) {
+                        Ok(()) => true,
+                        Err(_) => false,
+                    }
+                } else {
+                    false
+                }
+            }
+            Err(_) => false,
+        }
+    }
+}
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
 pub struct PrivateKey {
@@ -24,7 +71,7 @@ impl PrivateKey {
 
     pub fn sign_ed(&self, msg: &[u8]) -> Result<Vec<u8>, Error> {
         let sk = SecretKey::from_bytes(&self.key)?;
-        let pk = PublicKey::from(&sk);
+        let pk = ed25519_dalek::PublicKey::from(&sk);
         Ok(ExpandedSecretKey::from(&sk)
             .sign(msg, &pk)
             .as_ref()
